Update expected output files.
We're cutting off the last N days of results, with N varying for
different output files. When these tests were run for the first time,
the last day was cut off, but when it's run now, that day is included.
Ideally, results would not depend on system time at all, but until
that's the case, we'll just expect that tests are run after
2019-02-07.